+++ docs/Changelog.md
@@ -1,8 +1,14 @@
 # Changelog
 
+## v6.2.0
+* Add support for Android 4.2
+* Remove all usages of kotlin synthetics
+* Many version bumps
+* Added Hilt dep
+
 ## v6.1.0
 * Add support for Android 4.1 RC03
-* Breaking: Updated iconics and buganizer dependencies
+* Breaking: Updated iconics, buganizer, and spotless dependencies (major version bump)
 
 ## v6.0.0
 * Add support for Android 4.1.x